+++ api/api-bucket-grpc.go
@@ -14,13 +14,13 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package portal
+package api
 
 import (
 	"context"
 
+	pb "github.com/minio/m3/api/stubs"
 	cluster "github.com/minio/m3/cluster"
-	pb "github.com/minio/m3/portal/stubs"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
