+++ cmd/m3/login.go
@@ -25,7 +25,7 @@ import (
 	"time"
 
 	"github.com/minio/cli"
-	pb "github.com/minio/m3/portal/stubs"
+	pb "github.com/minio/m3/api/stubs"
 	"golang.org/x/crypto/ssh/terminal"
 )
