+++ cmd/m3/tenant-user-add.go
@@ -21,7 +21,7 @@ import (
 	"fmt"
 
 	"github.com/minio/cli"
-	pb "github.com/minio/m3/portal/stubs"
+	pb "github.com/minio/m3/api/stubs"
 )
 
 var (
